+++ key.go
@@ -1,0 +1,14 @@
+package micrologger
+
+import (
+	"fmt"
+
+	kitlog "github.com/go-kit/kit/log"
+	"github.com/go-stack/stack"
+)
+
+func newCallerFunc(skip int) kitlog.Valuer {
+	return func() interface{} {
+		return fmt.Sprintf("%+v", stack.Caller(5+skip))
+	}
+}